+++ libratom/models/entity.py
@@ -1,3 +1,5 @@
+# pylint: disable=too-few-public-methods,missing-docstring
+
 from sqlalchemy import Column, Integer, String
 from sqlalchemy.ext.declarative import declarative_base
